fix: update util
