+++ src/pymmcore_widgets/hcs/_plate_widget.py
@@ -1,8 +1,27 @@
 from __future__ import annotations
 
+from itertools import product
+from typing import TYPE_CHECKING
+
+from qtpy.QtCore import QRectF, Qt
+from qtpy.QtWidgets import (
+    QGraphicsScene,
+    QGraphicsView,
+    QWidget,
+)
+from useq import WellPlate
+
+from ._graphics_items import _WellGraphicsItem
+
+if TYPE_CHECKING:
+    from qtpy.QtGui import QBrush, QPen
+
+    from ._graphics_items import Well
+
+
 from typing import TYPE_CHECKING, NamedTuple
 
-from qtpy.QtCore import Qt, Signal
+from qtpy.QtCore import Signal
 from qtpy.QtGui import QBrush, QPen
 from qtpy.QtWidgets import (
     QComboBox,
@@ -10,15 +29,10 @@ from qtpy.QtWidgets import (
     QLabel,
     QPushButton,
     QVBoxLayout,
-    QWidget,
 )
 
 from ._plate_graphics_scene import _PlateGraphicsScene
-from ._util import (
-    PLATES,
-    _ResizingGraphicsView,
-    draw_plate,
-)
+from ._util import PLATES, _ResizingGraphicsView
 
 if TYPE_CHECKING:
     from useq import WellPlate
@@ -144,3 +158,73 @@ class _PlateSelectorWidget(QWidget):
 
         draw_plate(self.view, self.scene, PLATES[plate_name], brush=BRUSH, pen=PEN)
         self.valueChanged.emit(self.value())
+
+
+def draw_plate(
+    view: QGraphicsView,
+    scene: QGraphicsScene,
+    plate: WellPlate,
+    brush: QBrush | None,
+    pen: QPen | None,
+    opacity: float = 1.0,
+    text: bool = True,
+) -> None:
+    """Draw all wells of the plate in a QGraphicsScene."""
+    # setting a custom well size in scene px. Using 10 times the well size in mm
+    # gives a good resolution in the viewer.
+    well_size_x, well_size_y = plate.well_size
+    well_spacing_x, well_spacing_y = plate.well_spacing
+
+    scene.clear()
+
+    if not well_size_x or not well_size_y:
+        return
+
+    well_scene_size = well_size_x * 10
+
+    # calculate the width and height of the well in scene px
+    if well_size_x == well_size_y:
+        well_width = well_height = well_scene_size
+    elif well_size_x > well_size_y:
+        well_width = well_scene_size
+        # keep the ratio between well_size_x and well_size_y
+        well_height = int(well_scene_size * well_size_y / well_size_x)
+    else:
+        # keep the ratio between well_size_x and well_size_y
+        well_width = int(well_scene_size * well_size_x / well_size_y)
+        well_height = well_scene_size
+
+    # calculate the spacing between wells
+    dx = well_spacing_x - well_size_x if well_spacing_x else 0
+    dy = well_spacing_y - well_size_y if well_spacing_y else 0
+
+    # convert the spacing between wells in pixels
+    dx_px = dx * well_width / well_size_x if well_spacing_x else 0
+    dy_px = dy * well_height / well_size_y if well_spacing_y else 0
+
+    # the text size is the well_height of the well divided by 3
+    text_size = well_height / 3 if text else None
+
+    # draw the wells and place them in their correct row/column position
+    for row, col in product(range(plate.rows), range(plate.columns)):
+        _x = (well_width * col) + (dx_px * col)
+        _y = (well_height * row) + (dy_px * row)
+        rect = QRectF(_x, _y, well_width, well_height)
+        w = _WellGraphicsItem(rect, row, col, plate.circular_wells, text_size)
+        w.brush = brush
+        w.pen = pen
+        w.setOpacity(opacity)
+        scene.addItem(w)
+
+    # set the scene size
+    plate_width = (well_width * plate.columns) + (dx_px * (plate.columns - 1))
+    plate_height = (well_height * plate.rows) + (dy_px * (plate.rows - 1))
+
+    # add some offset to the scene rect to leave some space around the plate
+    offset = 20
+    scene.setSceneRect(
+        -offset, -offset, plate_width + (offset * 2), plate_height + (offset * 2)
+    )
+
+    # fit scene in view
+    view.fitInView(scene.sceneRect(), Qt.AspectRatioMode.KeepAspectRatio)
